Create list_of_classes struct

+++ SIMPLE_DATABASE/struct_item.h
@@ -1,9 +1,10 @@
+#include <stddef.h>
 #ifndef __STRUCT_ITEM_
 #define __STRUCT_ITEM_
 
 typedef struct tagITEM
 {
 	char lastname[20], firstname[20];
-	unsigned age;
+	size_t age;
 } ITEM;
 #endif
